Parallelization (#58)
Optional parallelization support through the `parallel` feature.
Perhaps it is best to provide different methods for serial /
parallelization for the poly and ring crates and avoid `cfg_iter*`. This
would allow the crate user to have more freedom to use either serial or
parallelized methods, while having the `parallel` feature set. The
linear algebra crate interface is set this way.